pyproject.toml
[project]
name = "tiny-agent-py"
version = "0.1.0"
description = "A hyper-minimal, litellm-based agent framework with sliding-window SQLite memory and MCP support"
authors = [{ name = "Author", email = "author@example.com" }]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tiny_agent"]

[tool.semantic_release]
version_toml = [
    "pyproject.toml:project.version"
]
version_variables = [
    "tiny_agent/__init__.py:__version__"
]
build_command = "uv run python -m build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.11",
]
